[TASK] Add klaro library via npm
The usage of the npm package enables us to better manage the version of
klaro in the future as well as to provide e.g. the SCSS files.

This patch should not change the current behavior or style in any way,
as the same version v0.5.32 of klaro is used.

Additionally, a gulp task is included to easily update the version.