[TASK] Compile ConsentApp.js with babel
This allows the use of the latest JavaScript, which is then compiled by
Babel so that older browsers are still supported.

The file was also renamed from `ConsentController.js` to `ConsentApp.js`
to better match the method name. It also received a light cleanup.

This is preparatory work for new features.