[FEATURE] Introduce service opt-in on demand
If the use of a service has not been consented to, visitors should have
the possibility of opting-in on demand.

This can be achieved easily by using a custom callback function in a
service combined with the new method `checkConsentOptIn` provided by the
ConsentApp.

Example callback function:
```
function consentGoogleRecaptcha(state, app) {
  var $element = $('.g-recaptcha').closest('form');
  var $parentElement = $element.parent();
  ConsentApp.checkConsentOptIn(app, state, $element, $parentElement);
}
```

In case the state of the app is false, the ConsentApp automatically
injects a container offering the possibility to opt-in once or
permanently. The displayed opt-in texts are configurable directly within
the service in the TYPO3 Backend.