[TASK] Create scss version of extension theme
The klaro theme shipped by the extension is now created in scss. The
build chain was extended to create a minified and non-minified css file.

In future, this theme will be build directly on top of the scss sources
of klaro, which also improves the maintainability especially regarding
updates of klaro.