[TASK] Refactor scss theme based on klaro theme
The theme provided by the extension now uses the theme of klaro as
foundation. This significantly reduces the maintenance effort for future
updates of klaro, since only the necessary styles have been modified.

Breaking changes :
1. The theme expects the constant 'noticeAsModal' to be true in order to
benefit from already existing styles of Klaro. Therefore, the default
value of the constant was changed.
2. The default theme (constant 'stylePrefix') was changed to `center`.
3. The styling of the cookie list plugin was mostly removed in order to
fit better into existing project styles. It will probably fully
removed in the future.
4. The style prefix `notice--center-floated` was removed.

Important:
The updated theme is not exactly identical to previous versions.
However, it is possible to add further customizations with a minimal
effort or by creating an own theme for klaro.

The theme of the extension is supposed to become even more general in
the future, so that it can be integrated into any project without major
customization.